+++ src/chains.ts
@@ -1,26 +1,9 @@
-/**
- * Substrate Chain Type
- */
-export interface SubstrateChain {
-  network: string
-  name: string
-  rpcUrls: [string, ...string[]]
-  ss58Prefix?: number
-  explorerUrls?: Partial<Record<SubstrateExplorer, string>>
-  testnet?: boolean
-  faucetUrls?: string[]
-}
-
-export enum SubstrateExplorer {
-  Subscan = 'subscan',
-  PolkadotJs = 'polkadotjs',
-}
+import { SubstrateChain, SubstrateExplorer } from '@/types'
 
 /**
- * Defined Substrate Chain Constants
+ * Local Development Network
  */
 
-/// Local Development Network
 export const development: SubstrateChain = {
   network: 'development',
   name: 'Local Development',
@@ -35,7 +18,9 @@ export const development: SubstrateChain = {
   faucetUrls: ['https://polkadot.js.org/apps/#/accounts?rpc=ws://127.0.0.1:9944'],
 }
 
-/// Testnets
+/**
+ * Live Testnets
+ */
 
 export const alephzeroTestnet: SubstrateChain = {
   network: 'alephzero-testnet',
@@ -53,8 +38,9 @@ export const alephzeroTestnet: SubstrateChain = {
 
 export const rococo: SubstrateChain = {
   network: 'rococo',
-  name: 'Rococo',
-  rpcUrls: ['wss://rococo-rpc.polkadot.io'],
+  name: 'Rococo Contracts Testnet',
+  ss58Prefix: 42,
+  rpcUrls: ['wss://rococo-contracts-rpc.polkadot.io'],
   explorerUrls: {
     [SubstrateExplorer.Subscan]: `https://rococo.subscan.io`,
   },
@@ -74,7 +60,80 @@ export const shibuya: SubstrateChain = {
   faucetUrls: ['https://portal.astar.network/#/shibuya-testnet/assets'],
 }
 
-/// Canary Networks (Kusama)
+export const t0rnTestnet: SubstrateChain = {
+  network: 't0rn',
+  name: 'T0rn Testnet',
+  ss58Prefix: 42,
+  rpcUrls: ['wss://ws.t0rn.io'],
+  explorerUrls: {
+    [SubstrateExplorer.PolkadotJs]: `https://polkadot.js.org/apps/?rpc=${encodeURIComponent(
+      'wss://ws.t0rn.io',
+    )}/#/explorer`,
+  },
+  testnet: true,
+  faucetUrls: ['https://faucet.t0rn.io'],
+}
+
+export const bitCountryAlphaTestnet: SubstrateChain = {
+  network: 'bitcountry-alpha-testnet',
+  name: 'Bit.Country Alpha Testnet',
+  ss58Prefix: 268,
+  rpcUrls: ['wss://alphanet-rpc-gcp.bit.country'],
+  explorerUrls: {
+    [SubstrateExplorer.PolkadotJs]: `https://polkadot.js.org/apps/?rpc=${encodeURIComponent(
+      'wss://alphanet-rpc-gcp.bit.country',
+    )}#/explorer`,
+  },
+  testnet: true,
+  faucetUrls: ['https://testnet.bit.country/p/wallet/balance'],
+}
+
+export const agungTestnet: SubstrateChain = {
+  network: 'agung-testnet',
+  name: 'Agung Testnet',
+  ss58Prefix: 42,
+  rpcUrls: ['wss://wss.agung.peaq.network'],
+  explorerUrls: {
+    [SubstrateExplorer.Subscan]: `https://agung-testnet.subscan.io/`,
+    [SubstrateExplorer.PolkadotJs]: `https://polkadot.js.org/apps/?rpc=${encodeURIComponent(
+      'wss://wss.agung.peaq.network',
+    )}/#/explorer`,
+  },
+  testnet: true,
+  faucetUrls: ['https://discord.com/channels/943486047625572392/963415793394143232'],
+}
+
+export const amplitudeTestnet: SubstrateChain = {
+  network: 'amplitude-testnet',
+  name: 'Amplitude Testnet',
+  ss58Prefix: 57,
+  rpcUrls: ['wss://pencol-roc-00.pendulumchain.tech'],
+  explorerUrls: {
+    [SubstrateExplorer.PolkadotJs]: `https://polkadot.js.org/apps/?rpc=${encodeURIComponent(
+      'wss://pencol-roc-00.pendulumchain.tech',
+    )}/#/explorer`,
+  },
+  testnet: true,
+  faucetUrls: [],
+}
+
+export const phalaPOC5Testnet: SubstrateChain = {
+  network: 'phala-PoC-5-testnet',
+  name: 'Phala PoC-5 Testnet',
+  ss58Prefix: 30,
+  rpcUrls: ['wss://poc5.phala.network/ws'],
+  explorerUrls: {
+    [SubstrateExplorer.PolkadotJs]: `https://polkadot.js.org/apps/?rpc=${encodeURIComponent(
+      'wss://poc5.phala.network/ws',
+    )}#/explorer`,
+  },
+  testnet: true,
+  faucetUrls: [],
+}
+
+/**
+ * Live Canary Networks
+ */
 
 export const shiden: SubstrateChain = {
   network: 'shiden',
@@ -86,7 +145,39 @@ export const shiden: SubstrateChain = {
   },
 }
 
-/// Mainnets
+export const amplitude: SubstrateChain = {
+  network: 'amplitude',
+  name: 'Amplitude',
+  ss58Prefix: 57,
+  rpcUrls: ['wss://rpc-amplitude.pendulumchain.tech', 'wss://amplitude-rpc.dwellir.com'],
+  explorerUrls: {
+    [SubstrateExplorer.PolkadotJs]: `https://polkadot.js.org/apps/?rpc=${encodeURIComponent(
+      'wss://rpc-amplitude.pendulumchain.tech',
+    )}#/explorer`,
+  },
+}
+
+export const khala: SubstrateChain = {
+  network: 'khala',
+  name: 'Khala',
+  ss58Prefix: 30,
+  rpcUrls: [
+    'wss://khala-api.phala.network/ws',
+    'wss://khala.api.onfinality.io/public-ws',
+    'wss://khala-rpc.dwellir.com',
+    'wss://public-rpc.pinknode.io/khala',
+  ],
+  explorerUrls: {
+    [SubstrateExplorer.PolkadotJs]: `https://polkadot.js.org/apps/?rpc=${encodeURIComponent(
+      'wss://khala-api.phala.network/ws',
+    )}#/explorer`,
+    [SubstrateExplorer.Subscan]: `https://khala.subscan.io`,
+  },
+}
+
+/**
+ * Live Mainnet Networks
+ */
 
 export const alephzero: SubstrateChain = {
   network: 'alephzero',
@@ -109,7 +200,31 @@ export const astar: SubstrateChain = {
   explorerUrls: {
     [SubstrateExplorer.Subscan]: `https://astar.subscan.io`,
   },
-  faucetUrls: [],
+}
+
+export const pendulum: SubstrateChain = {
+  network: 'pendulum',
+  name: 'Pendulum',
+  ss58Prefix: 56,
+  rpcUrls: ['wss://rpc-pendulum.prd.pendulumchain.tech', 'wss://pendulum-rpc.dwellir.com'],
+  explorerUrls: {
+    [SubstrateExplorer.PolkadotJs]: `https://polkadot.js.org/apps/?rpc=${encodeURIComponent(
+      'wss://rpc-pendulum.prd.pendulumchain.tech',
+    )}#/explorer`,
+  },
+}
+
+export const phala: SubstrateChain = {
+  network: 'phala',
+  name: 'Phala',
+  ss58Prefix: 30,
+  rpcUrls: ['wss://api.phala.network/ws', 'wss://phala.api.onfinality.io/public-ws'],
+  explorerUrls: {
+    [SubstrateExplorer.PolkadotJs]: `https://polkadot.js.org/apps/?rpc=${encodeURIComponent(
+      'wss://api.phala.network/ws',
+    )}#/explorer`,
+    [SubstrateExplorer.Subscan]: `https://phala.subscan.io`,
+  },
 }
 
 /**
@@ -119,10 +234,19 @@ export const allSubstrateChains: SubstrateChain[] = [
   development,
   alephzeroTestnet,
   rococo,
+  t0rnTestnet,
+  bitCountryAlphaTestnet,
+  agungTestnet,
+  amplitudeTestnet,
+  phalaPOC5Testnet,
   shibuya,
   shiden,
+  amplitude,
+  khala,
   alephzero,
   astar,
+  pendulum,
+  phala,
 ]
 
 /**