+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # @scio-labs/use-inkathon
 
+## 0.2.0
+
+### Minor Changes
+
+- [#36](https://github.com/scio-labs/use-inkathon/pull/36) [`a8e633d`](https://github.com/scio-labs/use-inkathon/commit/a8e633d95329d7b9661e0fa7f6e9a8154c020e5d) Thanks [@ical10](https://github.com/ical10)! - Added native support for more !ink compatible testnets (t0rn, Bit.Country, Peaq, Pendulum, Phala) and mainnets (Khala, Phala, Amplitude, Pendulum)
+
+## 0.1.1
+
+### Patch Changes
+
+- [#38](https://github.com/scio-labs/use-inkathon/pull/38) [`9542a0c`](https://github.com/scio-labs/use-inkathon/commit/9542a0c59e0e1dd9c7c626ce3a33f9c389737181) Thanks [@wottpal](https://github.com/wottpal)! - Add example packages (Vanilla React & Basic Scripts) to the workspace for development ease and demonstration purposes.
+
+- [#38](https://github.com/scio-labs/use-inkathon/pull/38) [`bd57d66`](https://github.com/scio-labs/use-inkathon/commit/bd57d66b61774dfa78af67003afd20c3b415f8a5) Thanks [@wottpal](https://github.com/wottpal)! - Fix the incorrect Rococo Contracts node RPC URL
+
 ## 0.1.0
 
 ### Minor Changes